+++ extra/prefix.js
@@ -9,6 +9,8 @@ var isImmutableClass = immutableClass.isImmutableClass;
 var immutableEqual = immutableClass.immutableEqual;
 var immutableArraysEqual = immutableClass.immutableArraysEqual;
 var immutableLookupsEqual = immutableClass.immutableLookupsEqual;
+var SimpleArray = immutableClass.SimpleArray;
+var NamedArray = immutableClass.NamedArray;
 
 var Chronoshift = require('chronoshift');
 var Timezone = Chronoshift.Timezone;